requirements.txt
async-timeout==4.0.2
attrs==21.4.0
click==8.0.4
cycler==0.11.0
dataclasses==0.6
Deprecated==1.2.13
filelock==3.4.1
grpcio==1.43.0
importlib-metadata==4.8.3
jsonschema==3.2.0
kiwisolver==1.3.1
matplotlib==3.3.4
msgpack==1.0.3
numpy==1.19.5
packaging==21.3
Pillow==8.4.0
protobuf==3.19.4
py-cpuinfo==8.0.0
pyparsing==3.0.7
pyrsistent==0.18.0
python-dateutil==2.8.2
PyYAML==6.0
ray==1.11.0
redis==4.2.1
scipy==1.5.4
six==1.16.0
typing_extensions==4.1.1
wrapt==1.14.0
zipp==3.6.0
tqdm==4.65.0
